+++ sdk/relationships/azure-resourcemanager-relationships/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Release History
 
-## 1.0.0-beta.2 (2026-08-13)
+## 1.0.0 (2026-09-03)
 
+- Azure Resource Manager relationships client library for Java. This package contains Microsoft Azure SDK for relationships Management SDK. Microsoft.Relationships Resource Provider management API. Package api-version 2026-08-01. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 - Azure Resource Manager relationships client library for Java. This package contains Microsoft Azure SDK for relationships Management SDK. Microsoft.Relationships Resource Provider management API. Package api-version 2026-03-01-preview. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 
 ### Breaking Changes
@@ -28,6 +29,8 @@
 
 * `models.ContainsRelationships` was added
 
+* `models.DependencyOfRelationshipsByServiceGroups` was added
+
 * `models.ServiceGroupMemberRelationshipPropertiesV2` was added
 
 #### `models.ServiceGroupMemberRelationship$Update` was modified
@@ -41,11 +44,12 @@
 
 #### `models.DependencyOfRelationships` was modified
 
-* `listByParent(java.lang.String)` was added
 * `listByParent(java.lang.String,com.azure.core.util.Context)` was added
+* `listByParent(java.lang.String)` was added
 
 #### `RelationshipsManager` was modified
 
+* `dependencyOfRelationshipsByServiceGroups()` was added
 * `containsRelationships()` was added
 
 #### `models.ServiceGroupMemberRelationship$Definition` was modified